refactor(auth): inline buildLegacyRequest to avoid widening type-check scope
Instead of moving buildLegacyRequest to packages/lib (which widened the TypeScript
dependency graph and caused type-check to evaluate more packages with pre-existing
Prisma type conflicts), inline a minimal helper directly in calcomHandler.ts.

The inlined helper uses structural types (HeadersLike, CookiesLike) instead of
importing from apps/web, avoiding cross-app dependencies while removing the
@calcom/web import.

This completes the removal of all 77 production imports of @calcom/web from
packages/* without widening the type-check scope.
