+++ packages/features/auth/signup/handlers/calcomHandler.ts
@@ -1,5 +1,6 @@
 import { cookies, headers } from "next/headers";
 import { NextResponse } from "next/server";
+import type { NextApiRequest } from "next";
 
 import { getPremiumMonthlyPlanPriceId } from "@calcom/app-store/stripepayment/lib/utils";
 import { getLocaleFromRequest } from "@calcom/features/auth/lib/getLocaleFromRequest";
@@ -20,7 +21,6 @@ import { prisma } from "@calcom/prisma";
 import { CreationSource } from "@calcom/prisma/enums";
 import { IdentityProvider } from "@calcom/prisma/enums";
 import { signupSchema } from "@calcom/prisma/zod-utils";
-import { buildLegacyRequest } from "@calcom/lib/buildLegacyCtx";
 
 import { joinAnyChildTeamOnOrgInvite } from "../utils/organization";
 import {
@@ -30,6 +30,33 @@ import {
 } from "../utils/token";
 
 const log = logger.getSubLogger({ prefix: ["signupCalcomHandler"] });
+
+type HeadersLike = { entries(): IterableIterator<[string, string]> };
+type CookiesLike = { getAll(): { name: string; value: string }[] };
+
+const createProxifiedObject = (object: Record<string, string>) =>
+  new Proxy(object, {
+    set: () => {
+      throw new Error("You are trying to modify 'headers' or 'cookies', which is not supported in app dir");
+    },
+  });
+
+const buildLegacyHeaders = (headers: HeadersLike) => {
+  const headersObject = Object.fromEntries(headers.entries());
+  return createProxifiedObject(headersObject);
+};
+
+const buildLegacyCookies = (cookies: CookiesLike) => {
+  const cookiesObject = cookies.getAll().reduce<Record<string, string>>((acc, { name, value }) => {
+    acc[name] = value;
+    return acc;
+  }, {});
+  return createProxifiedObject(cookiesObject);
+};
+
+const buildLegacyRequest = (headers: HeadersLike, cookies: CookiesLike) => {
+  return { headers: buildLegacyHeaders(headers), cookies: buildLegacyCookies(cookies) } as NextApiRequest;
+};
 
 const handler: CustomNextApiHandler = async (body, usernameStatus) => {
   const {
